Devel

+++ src/mission_publisher/include/mission_publisher.hpp
@@ -57,7 +57,7 @@ private:
     struct Mission {
         int drone_id;
         rclcpp::Time start_time;
-        std::vector<float> velocities;
+        std::vector<double> velocities;
 
         std::vector<geometry_msgs::msg::Pose> poses;
     };
